unet/requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
Pillow
tqdm